1.clear int and int64 transform warning. Give GetInt32 GetPropertyInt32 (still store int64) 2.GB2312 file to UTF8 3.typedef int to NFSOCK 4.NFDateTime GetTotalXXX return long long  instead of double 5.NFDateTime AddSeconds 6.all vcxproj rely on new NFCommon.props to be unified management 7.move tempDir to Temp\Debug_x64\projName\

+++ NFComm/NFCore/NFConsistentHash.hpp
@@ -193,7 +193,7 @@ public:
 
 	virtual bool GetSuitNodeRandom(NFCVirtualNode<T>& node)
 	{
-		int nID = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
+		int nID = (int) std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
 		return GetSuitNode(nID, node);
 	}
